+++ NEWS.md
@@ -1,3 +1,13 @@
+# CohortSurvival 1.1.1
+
+* Remove duplicate outcome cohort validation by @KimLopezGuell #395
+* Add weighted Kaplan-Meier support to survival estimation by @KimLopezGuell #393
+* Allow `censorOnDate` to be supplied as a cohort column in `addCohortSurvival()` by @KimLopezGuell #392
+* Move `CDMConnector` to suggests by @KimLopezGuell #391
+* Allow `type = NULL` in `plotSurvival()` by @KimLopezGuell #390
+* Add BugReports by @catalamarti #383
+
+
 # CohortSurvival 1.1.0
 
 * Use "first" option in addCohortSurvival to remove warning by @KimLopezGuell #375